minimal_requirements.txt
transformers==4.46.3
premsql==0.2.10
beautifulsoup4==4.12.3
fastapi==0.115.5
fastapi-cli==0.0.5
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-ollama==0.2.0
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
uvicorn==0.32.0
selenium==4.26.1
SQLAlchemy==2.0.35
numpy==1.26.4
ollama==0.3.3
openai==1.54.4
python-dotenv==1.0.1
letta==0.5.5
trafilatura==1.12.2
